feat: implement dashboard and login pages with user authentication and loading states
