+++ todo-app/app/login/page.tsx
@@ -1,19 +1,75 @@
-import { GalleryVerticalEnd } from "lucide-react"
+"use client";
 
-import { LoginForm } from "@/components/login-form"
+import { LoginForm } from "@/components/auth/LoginForm";
+import { ThemeToggle } from "@/components/ui/theme-toggle";
+import { useAuth } from "@/context/AuthContext";
+import { CheckSquare, Loader2 } from "lucide-react";
+import { useRouter } from "next/navigation";
+import { useEffect } from "react";
 
 export default function LoginPage() {
+  const { user, loading } = useAuth();
+  const router = useRouter();
+
+  // Redirigir a dashboard si ya está autenticado
+  useEffect(() => {
+    if (user && !loading) {
+      router.push("/dashboard");
+    }
+  }, [user, loading, router]);
+
+  // Mostrar loading mientras se verifica la autenticación
+  if (loading) {
+    return (
+      <div className="min-h-screen flex items-center justify-center">
+        <Loader2 className="h-8 w-8 animate-spin" />
+      </div>
+    );
+  }
+
+  // Si ya está autenticado, mostrar loading mientras redirige
+  if (user) {
+    return (
+      <div className="min-h-screen flex items-center justify-center">
+        <Loader2 className="h-8 w-8 animate-spin" />
+      </div>
+    );
+  }
+
   return (
-    <div className="bg-muted flex min-h-svh flex-col items-center justify-center gap-6 p-6 md:p-10">
-      <div className="flex w-full max-w-sm flex-col gap-6">
-        <a href="#" className="flex items-center gap-2 self-center font-medium">
-          <div className="bg-primary text-primary-foreground flex size-6 items-center justify-center rounded-md">
-            <GalleryVerticalEnd className="size-4" />
+    <div className="min-h-screen flex items-center justify-center bg-background p-4">
+      <div className="flex flex-col gap-6 w-full max-w-md relative">
+        {/* Theme Toggle */}
+        <div className="absolute top-0 right-0 z-10">
+          <ThemeToggle />
+        </div>
+
+        {/* Logo de la aplicación */}
+        <div className="text-center space-y-4">
+          <div className="w-16 h-16 bg-primary rounded-2xl flex items-center justify-center mx-auto">
+            <CheckSquare className="h-8 w-8 text-primary-foreground" />
           </div>
-          Acme Inc.
-        </a>
+          <div>
+            <h1 className="text-2xl font-bold text-foreground">NexToDo</h1>
+            <p className="text-muted-foreground">Organiza y gestiona tus tareas</p>
+          </div>
+        </div>
+
+        {/* Formulario de login */}
         <LoginForm />
+
+        <div className="text-muted-foreground text-center text-xs text-balance">
+          Al continuar, aceptas nuestros{" "}
+          <a href="#" className="underline underline-offset-4 hover:text-primary">
+            Términos de Servicio
+          </a>{" "}
+          y{" "}
+          <a href="#" className="underline underline-offset-4 hover:text-primary">
+            Política de Privacidad
+          </a>
+          .
+        </div>
       </div>
     </div>
-  )
+  );
 }
