--- todo-app/components/login-form.tsx
@@ -1,102 +1,0 @@
-"use client";
-
-import { Button } from "@/components/ui/button";
-import { Card, CardContent, CardDescription, CardHeader, CardTitle } from "@/components/ui/card";
-import { ThemeToggle } from "@/components/ui/theme-toggle";
-import { useAuth } from "@/context/AuthContext";
-import { cn } from "@/lib/utils";
-import { CheckSquare, Loader2 } from "lucide-react";
-
-export function LoginForm({ className, ...props }: React.ComponentProps<"div">) {
-  const { signInWithGoogle, signInAsGuest, loading } = useAuth();
-
-  return (
-    <div className={cn("flex flex-col gap-6 w-full max-w-md relative", className)} {...props}>
-      {/* Theme Toggle */}
-      <div className="absolute top-0 right-0 z-10">
-        <ThemeToggle />
-      </div>
-
-      {/* Logo de la aplicación */}
-      <div className="text-center space-y-4">
-        <div className="w-16 h-16 bg-primary rounded-2xl flex items-center justify-center mx-auto">
-          <CheckSquare className="h-8 w-8 text-primary-foreground" />
-        </div>
-        <div>
-          <h1 className="text-2xl font-bold text-foreground">NexToDo</h1>
-          <p className="text-muted-foreground">Organiza y gestiona tus tareas</p>
-        </div>
-      </div>
-
-      <Card className="shadow-sm">
-        <CardHeader className="text-center space-y-2">
-          <CardTitle className="text-xl">Bienvenido</CardTitle>
-          <CardDescription>Elige cómo quieres usar NexToDo</CardDescription>
-        </CardHeader>
-        <CardContent>
-          <div className="grid gap-3">
-            <Button
-              variant="default"
-              className="w-full h-11"
-              onClick={signInWithGoogle}
-              disabled={loading}
-            >
-              {loading ? (
-                <Loader2 className="mr-2 h-4 w-4 animate-spin" />
-              ) : (
-                <svg
-                  xmlns="http://www.w3.org/2000/svg"
-                  viewBox="0 0 24 24"
-                  className="mr-2 h-4 w-4"
-                >
-                  <path
-                    d="M12.48 10.92v3.28h7.84c-.24 1.84-.853 3.187-1.787 4.133-1.147 1.147-2.933 2.4-6.053 2.4-4.827 0-8.6-3.893-8.6-8.72s3.773-8.72 8.6-8.72c2.6 0 4.507 1.027 5.907 2.347l2.307-2.307C18.747 1.44 16.133 0 12.48 0 5.867 0 .307 5.387.307 12s5.56 12 12.173 12c3.573 0 6.267-1.173 8.373-3.36 2.16-2.16 2.84-5.213 2.84-7.667 0-.76-.053-1.467-.173-2.053H12.48z"
-                    fill="currentColor"
-                  />
-                </svg>
-              )}
-              {loading ? "Iniciando sesión..." : "Continuar con Google"}
-            </Button>
-
-            <div className="relative">
-              <div className="absolute inset-0 flex items-center">
-                <span className="w-full border-t" />
-              </div>
-              <div className="relative flex justify-center text-xs uppercase">
-                <span className="bg-background px-2 text-muted-foreground">o</span>
-              </div>
-            </div>
-
-            <Button
-              variant="outline"
-              className="w-full h-11"
-              onClick={signInAsGuest}
-              disabled={loading}
-            >
-              <CheckSquare className="mr-2 h-4 w-4" />
-              Demo local
-            </Button>
-
-            <p className="text-xs text-muted-foreground text-center">
-              <strong>Con Google:</strong> Datos sincronizados en la nube
-              <br />
-              <strong>Demo local:</strong> Datos guardados solo en este dispositivo
-            </p>
-          </div>
-        </CardContent>
-      </Card>
-
-      <div className="text-muted-foreground text-center text-xs text-balance">
-        Al continuar, aceptas nuestros{" "}
-        <a href="#" className="underline underline-offset-4 hover:text-primary">
-          Términos de Servicio
-        </a>{" "}
-        y{" "}
-        <a href="#" className="underline underline-offset-4 hover:text-primary">
-          Política de Privacidad
-        </a>
-        .
-      </div>
-    </div>
-  );
-}
